Interactive GraphPoint introducing
2 seperate class
